new Windows interpolation knocks out 1ms jitter in previous
remove dead code

thin #ifdef SYS_WINNT forest

64-bit time_t truncation fixes, warnings silenced

do not set hardware clock on exit unless system is shutting down

+++ ntpd/ntp_crypto.c
@@ -1934,7 +1934,7 @@ asn2ntp	(
 	tm.tm_wday = 0;
 	tm.tm_yday = 0;
 	tm.tm_isdst = 0;
-	return (timegm(&tm) + JAN_1970);
+	return ((u_long)timegm(&tm) + JAN_1970);
 }
